Virtual Threads demo
We can upgrade our app to Java 21 and benefit from virtual threads by
opting-in to it using a single property. Spring Boot will adapt the
infrastructure it autoconfigures to be Virtual Threads friendly.

We can see the impact with JConsole where our threading for Tomcat and
the scheduling pool are replaced by green threads.