Basic Testcontainers demo
This commit uses Testcontainers for the test so that a database should
not be available for it to succeed. We use `@DynamicPropertySource` to
inject the credentials exposed by the container so that Spring Boot is
configured to use it.

This also removes any reference of H2 and our duplicate management of
the schema.